# docs/swagger.yaml
basePath: /api
definitions:
...
  APIError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  APIErrors:
    properties:
      code:
        type: integer
      errors:
        items:
          $ref: '#/definitions/APIError'
        type: array
    type: object
...
  Blockchain:
    enum:
    - arbitrum
    - bitcoin
    - bitcoincash
    - bsc
    - dogecoin
    - ethereum
    - litecoin
    - polygon
    - tron
    type: string
    x-enum-varnames:
    - BlockchainArbitrum
    - BlockchainBitcoin
    - BlockchainBitcoinCash
    - BlockchainBinanceSmartChain
    - BlockchainDogecoin
    - BlockchainEthereum
    - BlockchainLitecoin
    - BlockchainPolygon
    - BlockchainTron
...
  FullPagingData:
    properties:
      last_page:
        type: integer
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  JSONResponse-ResponseWithFullPagination-github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions_FindRow:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ResponseWithFullPagination-github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions_FindRow'
      message:
        type: string
    type: object
...
  JSONResponse-TransactionInfoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/TransactionInfoResponse'
      message:
        type: string
    type: object
...
  JSONResponse-array_StoreTransactionResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/StoreTransactionResponse'
        type: array
      message:
        type: string
    type: object
...
  JSONResponse-github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions_StatisticsRow:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions.StatisticsRow'
      message:
        type: string
    type: object
...
  JSONResponse-string:
    properties:
      code:
        type: integer
      data:
        type: string
      message:
        type: string
    type: object
...
  ResponseWithFullPagination-github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions_FindRow:
    properties:
      items:
        items:
          $ref: '#/definitions/github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions.FindRow'
        type: array
      pagination:
        $ref: '#/definitions/FullPagingData'
    type: object
...
  StoreTransactionResponse:
    properties:
      amount:
        type: number
      amount_usd:
        type: number
      bc_uniq_key:
        type: string
      blockchain:
        type: string
      currency_id:
        type: string
      fee:
        type: number
      from_address:
        type: string
      id:
        format: uuid
        type: string
      network_created_at:
        format: date-time
        type: string
      receipt_id:
        format: uuid
        type: string
      store_id:
        format: uuid
        type: string
      to_address:
        type: string
      tx_hash:
        type: string
      type:
        $ref: '#/definitions/TransactionsType'
      user_id:
        format: uuid
        type: string
      wallet_id:
        format: uuid
        type: string
      withdrawal_is_manual:
        type: boolean
    type: object
...
  TransactionInfoResponse:
    properties:
      amount:
        type: number
      amount_usd:
        type: number
      bc_uniq_key:
        type: string
      blockchain:
        type: string
      created_at:
        type: string
      currency_id:
        type: string
      fee:
        type: number
      from_address:
        type: string
      id:
        type: string
      is_confirmed:
        type: boolean
      network_created_at:
        type: string
      receipt_id:
        type: string
      store_id:
        type: string
      to_address:
        type: string
      tx_hash:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      wallet:
        $ref: '#/definitions/WalletInfoResponse'
      webhook_history:
        items:
          $ref: '#/definitions/WhSendHistoryResponse'
        type: array
      withdrawal_is_manual:
        type: boolean
    type: object
...
  TransactionsType:
    enum:
    - transfer
    - deposit
    - withdrawal_from_processing
    type: string
    x-enum-varnames:
    - TransactionsTypeTransfer
    - TransactionsTypeDeposit
    - TransactionsTypeWithdrawalFromProcessing
...
  WalletInfoResponse:
    properties:
      id:
        type: string
      store_external_id:
        type: string
      wallet_created_at:
        type: string
      wallet_store_id:
        type: string
      wallet_updated_at:
        type: string
    type: object
...
  WhSendHistoryResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_success:
        type: boolean
      request:
        type: string
      response:
        type: string
      status_code:
        type: integer
      store_id:
        type: string
      url:
        type: string
      wh_type:
        type: string
    type: object
...
  github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions.FindRow:
    properties:
      amount:
        type: number
      amount_usd:
        type: number
      bc_uniq_key:
        type: string
      blockchain:
        type: string
      created_at:
        type: string
      created_at_index:
        $ref: '#/definitions/pgtype.Int8'
      currency:
        properties:
          blockchain:
            $ref: '#/definitions/Blockchain'
          code:
            type: string
          id:
            type: string
          name:
            type: string
        type: object
      currency_id:
        type: string
      fee:
        type: number
      from_address:
        type: string
      id:
        type: string
      is_system:
        type: boolean
      network_created_at:
        type: string
      receipt_id:
        type: string
      store_id:
        type: string
      to_address:
        type: string
      tx_hash:
        type: string
      type:
        $ref: '#/definitions/TransactionsType'
      untrusted_email:
        type: string
      updated_at:
        type: string
      user_email:
        type: string
      user_id:
        type: string
      wallet_id:
        type: string
      withdrawal_is_manual:
        type: boolean
    type: object
  github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions.Resolution:
    enum:
    - hour
    - day
    type: string
    x-enum-varnames:
    - ResolutionHour
    - ResolutionDay
  github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions.StatisticsRow:
    properties:
      amount_usd:
        type: number
      currency_ids:
        items:
          type: string
        type: array
      date:
        type: string
      resolution:
        $ref: '#/definitions/github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions.Resolution'
    type: object
...
  pgtype.Int8:
    properties:
      int64:
        type: integer
      valid:
        type: boolean
    type: object
info:
  contact:
    email: support@dv.net
    name: DV Support
  description: This is an API for DV Merchant
  title: DV Merchant
  version: "1.0"
paths:
...
  /v1/dv-admin/store/{id}/transactions/:
    get:
      consumes:
      - application/json
      description: Load store transactions
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-array_StoreTransactionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Load store transactions
      tags:
      - Store
...
  /v1/dv-admin/transaction/:
    get:
      consumes:
      - application/json
      description: Load user transactions
      parameters:
      - enum:
        - arbitrum
        - bitcoin
        - bitcoincash
        - bsc
        - dogecoin
        - ethereum
        - litecoin
        - polygon
        - tron
        in: query
        name: blockchain
        type: string
        x-enum-varnames:
        - BlockchainArbitrum
        - BlockchainBitcoin
        - BlockchainBitcoinCash
        - BlockchainBinanceSmartChain
        - BlockchainDogecoin
        - BlockchainEthereum
        - BlockchainLitecoin
        - BlockchainPolygon
        - BlockchainTron
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: currencies
        type: array
      - format: date-time
        in: query
        name: date_from
        type: string
      - format: date-time
        in: query
        name: date_to
        type: string
      - in: query
        name: from_address
        type: string
      - in: query
        name: is_system
        type: boolean
      - in: query
        name: min_amount_usd
        type: number
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - enum:
        - created_at_index
        - amount_usd
        - tx_hash
        - user_email
        in: query
        name: sort_by
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: sort_direction
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: store_uuids
        type: array
      - in: query
        name: to_address
        type: string
      - enum:
        - transfer
        - deposit
        - withdrawal_from_processing
        in: query
        name: type
        type: string
        x-enum-varnames:
        - TransactionsTypeTransfer
        - TransactionsTypeDeposit
        - TransactionsTypeWithdrawalFromProcessing
      - in: query
        name: wallet_address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ResponseWithFullPagination-github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions_FindRow'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Load user transactions
      tags:
      - Transaction
  /v1/dv-admin/transaction/{hash}:
    get:
      consumes:
      - application/json
      description: 'Search transactions by hash **Deprecated**: Use `/v1/dv-admin/search/:searchParam
        instead'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-TransactionInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Search transactions by hash
      tags:
      - Transaction
  /v1/dv-admin/transaction/{id}/send-webhooks:
    post:
      consumes:
      - application/json
      description: Send webhook manually. Accepts confirmed/unconfirmed tx
      parameters:
      - description: Transaction id
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Send webhook
      tags:
      - StoreWebhook
  /v1/dv-admin/transaction/export:
    get:
      consumes:
      - application/json
      description: Download user transactions
      parameters:
      - enum:
        - arbitrum
        - bitcoin
        - bitcoincash
        - bsc
        - dogecoin
        - ethereum
        - litecoin
        - polygon
        - tron
        in: query
        name: blockchain
        type: string
        x-enum-varnames:
        - BlockchainArbitrum
        - BlockchainBitcoin
        - BlockchainBitcoinCash
        - BlockchainBinanceSmartChain
        - BlockchainDogecoin
        - BlockchainEthereum
        - BlockchainLitecoin
        - BlockchainPolygon
        - BlockchainTron
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: currencies
        type: array
      - format: date-time
        in: query
        name: date_from
        type: string
      - format: date-time
        in: query
        name: date_to
        type: string
      - in: query
        name: from_address
        type: string
      - in: query
        name: is_system
        type: boolean
      - in: query
        name: min_amount_usd
        type: number
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - enum:
        - created_at_index
        - amount_usd
        - tx_hash
        - user_email
        in: query
        name: sort_by
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: sort_direction
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: store_uuids
        type: array
      - in: query
        name: to_address
        type: string
      - enum:
        - transfer
        - deposit
        - withdrawal_from_processing
        in: query
        name: type
        type: string
        x-enum-varnames:
        - TransactionsTypeTransfer
        - TransactionsTypeDeposit
        - TransactionsTypeWithdrawalFromProcessing
      - in: query
        name: wallet_address
        type: string
      produces:
      - application/json
      responses:
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Download user transactions
      tags:
      - Transaction
  /v1/dv-admin/transaction/stats:
    get:
      consumes:
      - application/json
      description: Load user transactions statistic by date
      parameters:
      - enum:
        - arbitrum
        - bitcoin
        - bitcoincash
        - bsc
        - dogecoin
        - ethereum
        - litecoin
        - polygon
        - tron
        in: query
        name: blockchain
        type: string
        x-enum-varnames:
        - BlockchainArbitrum
        - BlockchainBitcoin
        - BlockchainBitcoinCash
        - BlockchainBinanceSmartChain
        - BlockchainDogecoin
        - BlockchainEthereum
        - BlockchainLitecoin
        - BlockchainPolygon
        - BlockchainTron
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: currencies
        type: array
      - format: date-time
        in: query
        name: date_from
        type: string
      - format: date-time
        in: query
        name: date_to
        type: string
      - in: query
        name: is_system
        type: boolean
      - in: query
        name: min_amount_usd
        type: number
      - enum:
        - day
        - hour
        in: query
        name: resolution
        required: true
        type: string
      - enum:
        - transfer
        - deposit
        - withdrawal_from_processing
        in: query
        name: type
        type: string
        x-enum-varnames:
        - TransactionsTypeTransfer
        - TransactionsTypeDeposit
        - TransactionsTypeWithdrawalFromProcessing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-github.com/dv-net/dv-merchant_internal_storage_repos_repo_transactions_StatisticsRow'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Load user transactions statistic by date
      tags:
      - Transaction
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  XApiKey:
    description: Store API key
    in: header
    name: X-Api-Key
    type: apiKey
swagger: "2.0"